+++ generate/comments.js
@@ -2,12 +2,22 @@ const fs = require('fs');
 const chunkNLines = require('./helpers.js').chunkNLines;
 const calculateTime = require('./helpers.js').calculateTime;
 
+/*
+ *
+ * PURPOSE: to write (or overwrite) a .CSV
+ * file to be loaded into Postgres. Each CSV
+ * will correspond to a column identifier inside
+ * the table that the .CSV corresponds to.
+ * 
+ */
+
 let commentData = fs.createWriteStream('./csv_data/comments.csv');
 commentData.setMaxListeners(0);
 
 let data = 'message, user id, song id\n';
 
 let loadCommentData = () => {
+  console.log('Now loading: comments');
   let start = Date.now();
   commentData.write(data);
   let numOfWritesLeft = 100000000;
@@ -30,4 +40,4 @@ let loadCommentData = () => {
   writeComments();
 };
 
-loadCommentData();
+commentData.on('open', loadCommentData);